changes for the Purchase Options and Renwal Options for the Lease assets

+++ database/migrations/2019_01_01_085419_create_table_purchase_option.php
@@ -24,7 +24,7 @@ class CreateTablePurchaseOption extends Migration
             $table->date('expected_purchase_date')->nullable();
             $table->date('expected_lease_end_date')->nullable();
             $table->string('currency')->nullable();
-            $table->unsignedInteger('purchase_price')->nullable();
+            $table->decimal('purchase_price',12,2)->nullable();
             $table->timestamps();
         });
     }
